+++ src/contexts/data_context.provider.tsx
@@ -1,49 +1,10 @@
 import OCL from 'openchemlib/full';
 import { MoleculesDB } from 'openchemlib-utils';
-import {
-  createContext,
-  ReactNode,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react';
+import { ReactNode, useEffect, useMemo, useState } from 'react';
 
-import { ExtendedWikipediaJson, fetchData } from './fetch_data';
+import { ExtendedWikipediaJson, fetchData } from '../hooks/fetch_data';
 
-interface State {
-  allData: ExtendedWikipediaJson;
-  loading: boolean;
-  db: MoleculesDB;
-}
-
-const initData: ExtendedWikipediaJson = {
-  count: {
-    date: '',
-    molecules: 0,
-    errors: 0,
-    nogood: 0,
-    notfound: 0,
-    dup: 0,
-  },
-  data: {
-    molecules: [],
-    errors: [],
-    nogood: [],
-    notfound: [],
-    dup: [],
-  },
-};
-
-const DataContext = createContext<State>({
-  allData: initData,
-  loading: true,
-  db: new MoleculesDB(OCL),
-});
-
-export function useDataContext() {
-  return useContext(DataContext);
-}
+import { dataContext, initData } from './data_context';
 
 export function DataContextProvider({ children }: { children: ReactNode }) {
   const [allData, setAllData] = useState<ExtendedWikipediaJson>(initData);
@@ -70,5 +31,5 @@ export function DataContextProvider({ children }: { children: ReactNode }) {
     [allData, loading, db],
   );
 
-  return <DataContext.Provider value={value}>{children}</DataContext.Provider>;
+  return <dataContext.Provider value={value}>{children}</dataContext.Provider>;
 }
